+++ src/features/course/saveCourse.ts
@@ -1,5 +1,5 @@
 import { Course } from "./types";
-import { toStorage } from "../storage/save";
+import { toStorage } from "../storage";
 import { CoursesStorage } from "../../constant";
 
 export const saveCourses = (hostname: string, courses: Array<Course>): Promise<string> => {
