+++ src/features/quiz/getQuiz.ts
@@ -2,8 +2,7 @@ import { Quiz } from "./types";
 import { decodeQuizFromArray } from "./decode";
 import { Course } from "../course/types";
 import { fetchQuiz } from "../api/fetch";
-import { fromStorage } from "../storage/load";
-import { toStorage } from "../storage/save";
+import { toStorage, fromStorage } from "../storage";
 import { mergeEntities } from "../merge";
 import { QuizFetchTimeStorage, QuizzesStorage } from "../../constant";
