+++ src/features/setting/saveSetting.ts
@@ -1,5 +1,5 @@
 import { Settings } from "./types";
-import { toStorage } from "../storage/save";
+import { toStorage } from "../storage";
 import { SettingsStorage } from "../../constant";
 
 export const saveSettings = (hostname: string, settings: Settings): Promise<string> => {
